+++ openpkw-rest/src/main/java/org/openpkw/web/controllers/SessionsController.java
@@ -5,11 +5,12 @@ import org.openpkw.model.entity.UserDevice;
 import org.openpkw.qualifier.OpenPKWAPIController;
 import org.openpkw.repositories.UserDeviceRepository;
 import org.openpkw.repositories.UserRepository;
-import org.openpkw.web.dto.Token;
-import org.openpkw.web.dto.UserCredentialsDTO;
-import org.openpkw.web.validation.RequestValidator;
-import org.openpkw.web.validation.RestClientErrorMessage;
-import org.openpkw.web.validation.RestClientException;
+import org.openpkw.services.sessions.SessionService;
+import org.openpkw.services.sessions.dto.Token;
+import org.openpkw.services.user.dto.UserCredentialsDTO;
+import org.openpkw.validation.RequestValidator;
+import org.openpkw.validation.RestClientErrorMessage;
+import org.openpkw.validation.RestClientException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -29,59 +30,22 @@ import java.util.Optional;
 @RequestMapping("/sessions")
 public class SessionsController {
 
-    private final SecureRandom secureRandom = new SecureRandom();
-
-    @Inject
-    private UserRepository userRepository;
-
-    @Inject
-    private UserDeviceRepository deviceRepository;
-
     @Inject
     private RequestValidator registerUserValidator;
 
     @Inject
-    private PasswordEncoder passwordEncoder;
+    private SessionService sessionService;
 
     @RequestMapping(value = "/", method = RequestMethod.POST)
     public ResponseEntity<Map<String, String>> login(@RequestBody UserCredentialsDTO userCredentials) {
+        Token token ;
         try {
             registerUserValidator.validateUserAuthorization(userCredentials);
+            token = sessionService.login(userCredentials);
         } catch (RestClientException ex) {
             return buildResponse(ex.getErrorCode(), HttpStatus.BAD_REQUEST, null);
         }
-
-        Optional<User> user = userRepository.findByEmailAddress(userCredentials.getEmail());
-        if (!user.isPresent()) {
-            return buildResponse(RestClientErrorMessage.USER_NOT_FOUND, HttpStatus.BAD_REQUEST, null);
-        }
-
-        if (!passwordEncoder.matches(userCredentials.getPassword(), user.get().getPassword())) {
-            return buildResponse(RestClientErrorMessage.INVALID_PASSWORD, HttpStatus.BAD_REQUEST, null);
-        }
-
-        // W piewszej fazie implementacji zakładamy, że każdy użytkownik używa tylko jednego urządzenia
-        // Identyfikatorem urządzenia jest email użytkownika
-        String deviceID = userCredentials.getEmail();
-        UserDevice device = deviceRepository.findByUserIdAndDevId(user.get().getUserID(), deviceID);
-        if (device == null) {
-            device = new UserDevice();
-            device.setDevId(deviceID);
-            device.setUser(user.get());
-        }
-        Token token = createUserToken(device);
-        deviceRepository.saveAndFlush(device);
-
         return buildResponse(RestClientErrorMessage.OK, HttpStatus.OK, token.getToken());
-    }
-
-    private Token createUserToken(UserDevice device) {
-        byte[] bToken = new byte[24];
-        Token token = new Token();
-        secureRandom.nextBytes(bToken);
-        token.setToken(bToken);
-        device.setToken(token.getToken());
-        return token;
     }
 
     private ResponseEntity<Map<String, String>> buildResponse(RestClientErrorMessage validationError, HttpStatus httpStatus, String data) {
