+++ openpkw-rest/src/main/java/org/openpkw/web/controllers/UsersController.java
@@ -2,16 +2,16 @@ package org.openpkw.web.controllers;
 
 import org.openpkw.model.entity.User;
 import org.openpkw.model.entity.UserDevice;
-import org.openpkw.model.entity.UserType;
 import org.openpkw.qualifier.OpenPKWAPIController;
 import org.openpkw.repositories.UserDeviceRepository;
 import org.openpkw.repositories.UserRepository;
-import org.openpkw.web.dto.ResponseDTO;
-import org.openpkw.web.dto.UserDTO;
+import org.openpkw.services.user.UserService;
+import org.openpkw.services.user.dto.ResponseDTO;
+import org.openpkw.services.user.dto.UserDTO;
 import org.openpkw.web.helper.UserBuilder;
-import org.openpkw.web.validation.RequestValidator;
-import org.openpkw.web.validation.RestClientErrorMessage;
-import org.openpkw.web.validation.RestClientException;
+import org.openpkw.validation.RequestValidator;
+import org.openpkw.validation.RestClientErrorMessage;
+import org.openpkw.validation.RestClientException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -43,34 +43,17 @@ public class UsersController {
     @Inject
     private PasswordEncoder passwordEncoder;
 
+    @Inject
+    private UserService userService;
+
     @RequestMapping(value = "/", method = RequestMethod.POST, headers = "Accept=application/json")
-    @Transactional
     public ResponseEntity<Map<String, String>> register(@RequestBody UserDTO userRegister) {
         try {
             registerUserValidator.validateUserRegistration(userRegister);
+            userService.register(userRegister);
         } catch (RestClientException ex) {
             return buildResponse(ex.getErrorCode(), HttpStatus.BAD_REQUEST, null);
         }
-
-        Optional<User> userOptional = userRepository.findByEmailAddress(userRegister.getEmail());
-        if (userOptional.isPresent()) {
-            return buildResponse(RestClientErrorMessage.USER_ALREADY_EXISTS, HttpStatus.BAD_REQUEST, null);
-        }
-
-        User user = new User();
-        user.setEmail(userRegister.getEmail());
-        user.setPassword(passwordEncoder.encode(userRegister.getPassword()));
-        user.setFirstName(userRegister.getFirst_name());
-        user.setLastName(userRegister.getLast_name());
-        user.setIsActive(true);
-        user.setUserType(UserType.VOLUNTEER);
-
-        try {
-            userRepository.saveAndFlush(user);
-        } catch (Exception ex) {
-            throw new RuntimeException("Failed to register new user: " + ex.getMessage(), ex);
-        }
-
         return buildResponse(RestClientErrorMessage.OK, HttpStatus.OK, null);
     }
 
@@ -80,8 +63,7 @@ public class UsersController {
     @RequestMapping(value = "/{email:.+}", method = RequestMethod.GET, headers = "Accept=application/json")
     @Transactional
     public ResponseEntity<UserDTO> get(@PathVariable String email) throws InstantiationException, IllegalAccessException {
-        Optional<User> user = userRepository.findByEmailAddress(email);
-
+        Optional<User> user = userService.get(email);
         if (user.isPresent()) {
             return new ResponseEntity<>(userToJson(user.get()), HttpStatus.OK);
         } else {
@@ -94,15 +76,10 @@ public class UsersController {
     public ResponseEntity<Map<String, String>> delete(@PathVariable("email") String email) {
         Optional<User> user = userRepository.findByEmailAddress(email);
 
-        if (user != null) {
-            for (UserDevice device : user.get().getUserDevices()) {
-                deviceRepository.delete(device);
-            }
-            userRepository.delete(user.get());
+        if (user.isPresent())
             return buildResponse(RestClientErrorMessage.OK, HttpStatus.OK, null);
-        } else {
+        else
             return buildResponse(RestClientErrorMessage.USER_NOT_FOUND, HttpStatus.BAD_REQUEST, null);
-        }
     }
 
     private ResponseEntity<Map<String, String>> buildResponse(RestClientErrorMessage validationError, HttpStatus httpStatus, String data) {
@@ -112,6 +89,7 @@ public class UsersController {
         responseBody.put("data", data);
         return new ResponseEntity<Map<String, String>>(responseBody, httpStatus);
     }
+
 
     // TODO: Użyć jakiejś biblioteki
     private UserDTO userToJson(User user) {
