--- openpkw-rest/src/main/java/org/openpkw/web/dto/UserCredentialsDTO.java
@@ -1,27 +1,0 @@
-package org.openpkw.web.dto;
-
-import java.io.Serializable;
-
-public class UserCredentialsDTO implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private String email;
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
