+++ openpkw-services/pom.xml
@@ -101,6 +101,11 @@
 			<artifactId>opencsv</artifactId>
 			<version>2.3</version>
 		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-web</artifactId>
+		</dependency>
 	</dependencies>
 
 </project>
