+++ openpkw-services/src/main/java/org/openpkw/services/sessions/dto/Token.java
@@ -1,0 +1,24 @@
+package org.openpkw.services.sessions.dto;
+
+import java.util.Base64;
+
+public class Token {
+
+    private String token;
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public void setToken(byte[] bToken) {
+        token = Base64.getEncoder().encodeToString(bToken);
+    }
+
+    public void error() {
+        token = "ERROR";
+    }
+}
