+++ openpkw-services/src/main/java/org/openpkw/services/user/dto/ErrorDTO.java
@@ -1,4 +1,4 @@
-package org.openpkw.web.dto;
+package org.openpkw.services.user.dto;
 
 import java.io.Serializable;
