+++ openpkw-services/src/main/java/org/openpkw/validation/StringUtils.java
@@ -1,0 +1,10 @@
+package org.openpkw.validation;
+
+/**
+ * Created by mrozi on 14.01.16.
+ */
+public class StringUtils {
+    public static boolean isEmpty(String value){
+        return value==null || value.isEmpty();
+    }
+}
